% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_measles_data}
\alias{clean_measles_data}
\title{Clean Measles Data}
\usage{
clean_measles_data(data)
}
\arguments{
\item{data}{A data frame of raw measles data from CDC.}
}
\value{
A cleaned data frame.
}
\description{
This function cleans and formats the raw measles data from CDC.
}
\keyword{internal}
