% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_measles_metadata}
\alias{get_measles_metadata}
\title{Get Metadata for CDC Measles Dataset}
\usage{
get_measles_metadata()
}
\value{
A list containing metadata about the dataset.
}
\description{
This function returns metadata about the CDC measles dataset.
}
\examples{
\dontrun{
get_measles_metadata()
}

}
