% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_data_available}
\alias{is_data_available}
\title{Check if CDC Measles Data is Available}
\usage{
is_data_available(verbose = FALSE)
}
\arguments{
\item{verbose}{Logical. If TRUE, prints detailed messages during checks. Default is FALSE.}
}
\value{
Logical. TRUE if at least one data source is available, FALSE otherwise.
}
\description{
This function checks if the CDC measles data URLs are accessible.
}
\examples{
\dontrun{
is_data_available()
is_data_available(verbose = TRUE)
}

}
