% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_measles_state_map}
\alias{plot_measles_state_map}
\title{Create a Map of US Measles Cases by State}
\usage{
plot_measles_state_map(
  data,
  state_col = "state",
  cases_col = "cases",
  title = "Measles Cases by State"
)
}
\arguments{
\item{data}{A data frame containing measles data by state.}

\item{state_col}{Character. Name of the column containing state names/codes. Default is "state".}

\item{cases_col}{Character. Name of the column containing case counts. Default is "cases".}

\item{title}{Character. Plot title. Default is "Measles Cases by State".}
}
\value{
A ggplot object with the choropleth map.
}
\description{
This function creates a choropleth map of measles cases by US state.
}
\examples{
\dontrun{
data <- get_measles_data()
# Assuming data has state-level information
plot_measles_state_map(data)
}

}
