% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_measles_time_series}
\alias{plot_measles_time_series}
\title{Plot Measles Cases Over Time}
\usage{
plot_measles_time_series(
  data,
  date_col = "date",
  cases_col = "cases",
  title = "Measles Cases Over Time"
)
}
\arguments{
\item{data}{A data frame containing measles data, typically from get_measles_data().}

\item{date_col}{Character. Name of the column containing dates. Default is "date".}

\item{cases_col}{Character. Name of the column containing case counts. Default is "cases".}

\item{title}{Character. Plot title. Default is "Measles Cases Over Time".}
}
\value{
A ggplot object with the time series plot.
}
\description{
This function creates a time series plot of measles cases.
}
\examples{
\dontrun{
data <- get_measles_data()
plot_measles_time_series(data)
}

}
