% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_measles_data.R
\name{get_measles_data}
\alias{get_measles_data}
\title{Download CDC Measles Data}
\usage{
get_measles_data(save_file = FALSE, file_name = "cdc_measles_data.csv")
}
\arguments{
\item{save_file}{Logical. If TRUE, saves the downloaded data as a CSV file
in the current working directory. Default is FALSE.}

\item{file_name}{Character. The name of the file to save if save_file is TRUE.
Default is "cdc_measles_data.csv".}
}
\value{
A data frame containing CDC measles data.
}
\description{
This function downloads measles data from the CDC website.
The data includes reported measles cases across the United States.
}
\examples{
\dontrun{
measles_data <- get_measles_data()
measles_data <- get_measles_data(save_file = TRUE)
}

}
